[FIX] project: ensure the quick create bar persists after refresh
To reproduce the issue:
- In the project app, click on a project to enter the Kanban view
- Refresh the page

The "+ stage" button to add a stage to a project disappears upon page refresh.

This is caused by a context check that functions correctly during the normal flow. However, when directly accessing the
link or refreshing the page, the `active_model` context is lost. Since the action `act_project_project_2_project_task_all`
always filter project on active_id, we can add as a default context ´'active_model': 'project.project'´

opw-4075290

closes odoo/odoo#177791

X-original-commit: 29a866f968d6e64e57a0d8bbadd5ca16a4a92f56
Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>